requirements.txt
nnunetv2>=2.5
torch>=2.1
numpy>=1.24
scipy>=1.10
nibabel>=5.0
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0
tqdm>=4.65
SimpleITK>=2.2
